added more tests - finsihed testing get routes
